+++ src/pyblu/__init__.py
@@ -1,6 +1,6 @@
 """A Python library for controlling BluOS players."""
 
-from ._player import Player
-from ._entities import Status, Volume, SyncStatus, PairedPlayer, PlayQueue, Preset, Input
+from .player import Player
+from .entities import Status, Volume, SyncStatus, PairedPlayer, PlayQueue, Preset, Input
 
 __all__ = ["Player", "Status", "Volume", "SyncStatus", "PairedPlayer", "PlayQueue", "Preset", "Input"]
